+++ clients/core-java/src/main/java/org/commonjava/indy/client/core/metric/ClientMetrics.java
@@ -95,7 +95,7 @@ public class ClientMetrics
 
     public void registerEnd( HttpResponse response )
     {
-        if ( !enabled )
+        if ( !enabled || response == null )
         {
             return;
         }
