+++ pom.xml
@@ -63,6 +63,8 @@
     <logback-version>1.2.3</logback-version>
     <infinispanVersion>9.4.15.Final</infinispanVersion>
     <!--<luceneVersion>7.3.0</luceneVersion>-->
+    <protostreamVersion>4.3.0.Final</protostreamVersion>
+    <gsonVersion>2.8.2</gsonVersion>
     <hibernateVersion>5.8.1.Final</hibernateVersion>
     <resteasyVersion>3.0.12.Final</resteasyVersion>
     <undertowVersion>2.1.0.Final</undertowVersion>
@@ -82,7 +84,7 @@
     <cassandraUnitVersion>3.7.1.0</cassandraUnitVersion>
     <datastaxVersion>3.7.2</datastaxVersion>
     <pathmappedStorageVersion>1.8</pathmappedStorageVersion>
-    <o11yphantVersion>1.3</o11yphantVersion>
+    <o11yphantVersion>1.4-SNAPSHOT</o11yphantVersion>
 
     <!-- commonjava/redhat projects -->
     <atlasVersion>1.1.1</atlasVersion>
