+++ subsys/cassandra/pom.xml
@@ -34,6 +34,10 @@
             <groupId>org.commonjava.indy</groupId>
             <artifactId>indy-api</artifactId>
         </dependency>
+        <dependency>
+          <groupId>org.commonjava.indy</groupId>
+          <artifactId>indy-subsys-trace</artifactId>
+        </dependency>
         <!--Datastax Java Driver-->
         <dependency>
             <groupId>com.datastax.cassandra</groupId>
